+++ Tests/Unit/Configuration/WidgetsProviderTest.php
@@ -37,7 +37,7 @@ class WidgetsProviderTest extends TestCase
         self::assertContains('actionsPerDay', $actual);
         self::assertContains('actionsPerMonth', $actual);
         self::assertContains('annotations', $actual);
-//        self::assertContains('createAnnotation', $actual);
+        self::assertContains('createAnnotation', $actual);
         self::assertContains('bounceRate', $actual);
         self::assertContains('browsers', $actual);
         self::assertContains('campaigns', $actual);
@@ -62,7 +62,7 @@ class WidgetsProviderTest extends TestCase
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.actionsPerDay.title', 'actionsPerDay'], $actual);
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.actionsPerMonth.title', 'actionsPerMonth'], $actual);
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.annotations.title', 'annotations'], $actual);
-//        self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.createAnnotation.title', 'createAnnotation'], $actual);
+        self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.createAnnotation.title', 'createAnnotation'], $actual);
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.visitsSummary.bounceRate.title', 'bounceRate'], $actual);
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.referrers.campaigns.title', 'campaigns'], $actual);
         self::assertContains([Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.contents.contentNames.title', 'contentNames'], $actual);