+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Compatibility with Symfony 6
+- Widget "Create annotation" (#25)
 
 ## [1.2.0] - 2021-10-20